REFACTOR: put styles/theme default back to component file
