+++ src/components/Avatar/Avatar.component.jsx
@@ -2,7 +2,15 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import MaterialAvatar from '@material-ui/core/Avatar';
 
-import useStyles from './Avatar.styles';
+import makeStyles from '@material-ui/core/styles/makeStyles';
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    width: theme.spacing(20),
+    height: theme.spacing(20),
+    margin: [[0, 'auto']],
+  },
+}));
 
 function Avatar({ id, avatar, userName, ...props }) {
   const classes = useStyles({ classes: props.classes });
