--- src/components/Avatar/Avatar.styles.js
@@ -1,11 +1,0 @@
-import makeStyles from '@material-ui/core/styles/makeStyles';
-
-const useStyles = makeStyles(theme => ({
-  root: {
-    width: theme.spacing(20),
-    height: theme.spacing(20),
-    margin: [[0, 'auto']],
-  },
-}));
-
-export default makeStyles;
