+++ src/components/Card/Card.component.jsx
@@ -9,8 +9,29 @@ import {
   Link,
   Icon,
 } from '@material-ui/core';
+import makeStyles from '@material-ui/core/styles/makeStyles';
 
-import useStyles from './Card.styles';
+const useStyles = makeStyles(theme => ({
+  card: {
+    maxWidth: 300,
+    minHeight: 300,
+    textAlign: 'center',
+    position: 'relative',
+  },
+  cardMedia: {
+    minHeight: 300,
+  },
+  linkToSolution: {
+    border: '1px solid orange',
+    display: 'inline-block',
+    padding: '1rem',
+    borderRadius: '0.3em',
+  },
+  cardBottom: {
+    background: theme.palette.common.white,
+    color: theme.palette.common.black,
+  },
+}));
 
 /**
  * Card component. Accepts a few params:
